version 0.2
git-svn-id: https://plugins.svn.wordpress.org/alti-watermark/trunk@898196 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ index.php
@@ -1,12 +1,19 @@
 <?php 
 /*
-    Plugin Name: Watermark
-    Plugin URI: http://www.alticreation.com/en/alti-watermark/
-    Description: Add a watermark on all your pictures even the one which are already uploaded. You can setup this plugin through the Media category in the side menu.
-    Author: Alexis Blondin
-    Version: 0.1
-    Author URI: http://www.alticreation.com
+Plugin Name: Watermark
+Plugin URI: http://www.alticreation.com/en/alti-watermark/
+Description: Add a watermark on all your pictures even the one which are already uploaded. You can setup this plugin through the Media category in the side menu.
+Author: Alexis Blondin
+Version: 0.1
+Author URI: http://www.alticreation.com
 */
+
+// Translation
+function altiwatermark_i18n()
+{
+load_plugin_textdomain('altiwatermark', false, basename( dirname( __FILE__ ) ) . '/languages' );
+}
+add_action('plugins_loaded', 'altiwatermark_i18n');
 
 require_once('paths.php');
 
@@ -25,9 +32,25 @@ function altiwatermark_deactivate() {
 
   global $acw_uploads;
   if( file_exists($acw_uploads['basedir'].'/'.'.htaccess') ) {
-    
-    unlink( $acw_uploads['basedir'].'/'.'.htaccess' );
-    
+    preg_match('/^(.+)\# BEGIN alti\-watermark/s', file_get_contents( $acw_uploads['basedir'].'/'.'.htaccess' ), $htaccessContentPreviousContent);
+    if($htaccessContentPreviousContent[1] != '')  {
+      file_put_contents( $acw_uploads['basedir'].'/'.'.htaccess', $htaccessContentPreviousContent[1] );
+    } else {
+      unlink( $acw_uploads['basedir'].'/'.'.htaccess' );
+    }
+
+
   }
-  
+
 }
+
+// Setting link
+function altiwatermark_settings_link($links) { 
+  global $acw_plugins;
+  $settings_link = '<a href="'.get_admin_url().'upload.php?page='.$acw_plugins['subdir'].'/admin-page.php">'.__("Settings").'</a>'; 
+  array_unshift($links, $settings_link); 
+  return $links; 
+}
+
+$plugin = plugin_basename(__FILE__); 
+add_filter("plugin_action_links_$plugin", 'altiwatermark_settings_link' );
